+++ src/main/resources/application.conf
@@ -1,11 +1,5 @@
-es-priority-mailbox {
-  mailbox-type = "de.upb.cs.swt.delphi.webapi.ElasticPriorityMailbox"
-}
-
 akka.actor.deployment {
-  /espriomailboxactor {
-    mailbox = es-priority-mailbox
-  }
+
 }
 
 akka {
@@ -14,15 +8,4 @@ akka {
       remote-address-header = on
     }
   }
-}
-
-# Use this dispatcher for actors that make blocking calls to the Elasticsearch database
-elasticsearch-handling-dispatcher {
-  type = Dispatcher
-  executor = "thread-pool-executor"
-  thread-pool-executor {
-    fixed-pool-size = 4
-    # This thread pool is intended for development purposes, and should be increased for production
-  }
-  throughput = 1
 }
