--- src/main/scala/de/upb/cs/swt/delphi/webapi/ElasticPriorityMailbox.scala
@@ -1,30 +1,0 @@
-// Copyright (C) 2018 The Delphi Team.
-// See the LICENCE file distributed with this work for additional
-// information regarding copyright ownership.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package de.upb.cs.swt.delphi.webapi
-
-import akka.actor.ActorSystem
-import akka.dispatch.{PriorityGenerator, UnboundedStablePriorityMailbox}
-import de.upb.cs.swt.delphi.webapi.ElasticActorManager.{Enqueue, Retrieve}
-import com.typesafe.config.Config
-
-class ElasticPriorityMailbox (settings: ActorSystem.Settings, config: Config)
-  extends UnboundedStablePriorityMailbox(
-    PriorityGenerator{
-      case Retrieve(_) => 5
-      case Enqueue(_) => 1
-      case _ => 2
-    })
