+++ src/main/scala/de/upb/cs/swt/delphi/webapi/search/SearchQuery.scala
