+++ src/main/scala/de/upb/cs/swt/delphi/webapi/search/querylanguage/Syntax.scala
@@ -14,7 +14,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package de.upb.cs.swt.delphi.webapi.querylanguage
+package de.upb.cs.swt.delphi.webapi.search.querylanguage
 
 import org.parboiled2.{CharPredicate, Parser, ParserInput, Rule1}
